requirements.txt
django-mathfilters==0.3.0
Django==1.7.1
MySQL-python==1.2.5
South==1.0.2
babel==1.3
boto==2.34.0
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-classy-tags==0.5.2
django-compressor==1.4
django-extra-views==0.6.4
django-haystack==2.3.1
django-oscar==1.0.1
django-polymorphic==0.6
django-shop==0.2.0
django-shop-configurableproduct==0.3.1
django-storages==1.1.8
django-tables2==0.15.0
django-toolbelt==0.0.1
django-treebeard==2.0
djangorestframework==3.0.0
factory-boy==2.4.1
gunicorn==19.1.1
mock==1.0.1
phonenumbers==6.3.0
pillow==2.4.0
psycopg2==2.5.4
purl==1.0.3
pytz==2014.10
six==1.8.0
sorl-thumbnail==11.12.1b
static3==0.5.1
unidecode==0.04.17
virtualenv==1.11.6
wsgiref==0.1.2
django-sql-explorer==0.7
jsonfield==1.0.0
